+++ server/server.js
@@ -9,6 +9,7 @@ const port = 8510;
 
 const rootPath = path.join(__dirname, "../");
 const dataFolderPath = path.join(rootPath, "data");
+const autosavesFolderPath = path.join(dataFolderPath, "autosaves");
 const notebooksFolderPath = path.join(dataFolderPath, "notebooks");
 const imageFolderPath = path.join(dataFolderPath, "images");
 const attachmentsFolderPath = path.join(dataFolderPath, "attachments");
@@ -16,6 +17,7 @@ const masterFilePath = path.join(dataFolderPath, "master.json");
 
 const userDataFolders = [
   { name: "Data", path: dataFolderPath },
+  { name: "Autosaves", path: autosavesFolderPath },
   { name: "Notebooks", path: notebooksFolderPath },
   { name: "Image", path: imageFolderPath },
   { name: "Attachments", path: attachmentsFolderPath },
@@ -26,7 +28,7 @@ for (let i = 0; i < userDataFolders.length; i++) {
     fs.mkdirSync(userDataFolders[i].path);
     console.warn(`Created missing ${userDataFolders[i].name} Folder.`);
     console.log(
-      "This is standard if you've freshly cloned the Repository, as the data folder is ignored by git.",
+      "This is standard if you've freshly cloned the Repository or updated to a newer version, as the data folder is ignored by git.",
     );
   }
 }
@@ -35,7 +37,7 @@ for (let i = 0; i < userDataFolders.length; i++) {
 if (!fs.existsSync(masterFilePath)) {
   const masterFileContent = `[
   {
-    "usedImages": []
+    "unsavedFiles": []
   }
 ]
 `;
@@ -44,15 +46,63 @@ if (!fs.existsSync(masterFilePath)) {
   console.log(
     "This is standard if you've freshly cloned the Repository, as the data folder is ignored by git.",
   );
+} else {
+  deleteDeprecatedMasterProperties();
 }
 
-const exportRoute = require("./routes/export");
+// Deletes deprecated master.json properties
+async function deleteDeprecatedMasterProperties() {
+  console.log("Checking for deprecated master.json properties...");
+
+  let changesMade = false;
+  // Getting the masterfile data. Has to be parsed.
+  const rawMasterFile = await fs.readFileSync(masterFilePath, "utf-8");
+  const masterFile = JSON.parse(rawMasterFile);
+
+  // Deletes the usedImages property
+  if (masterFile[0].usedImages) {
+    delete masterFile[0].usedImages;
+    console.log('Deleted masterfile property "usedImages".');
+    changesMade = true;
+  }
+
+  if (changesMade) {
+    try {
+      // Updates the masterfile
+      fs.writeFileSync(masterFilePath, JSON.stringify(masterFile), "utf-8");
+      console.log("Successfully removed deprecated masterfile properties.");
+    } catch (err) {
+      console.error("Something went wrong trying to update master.json.");
+      console.error(err);
+    }
+  } else {
+    console.log(
+      "No deprecated masterfile properties found. Starting App without changes.",
+    );
+  }
+}
+
+app.get("/api/getMaster", async (req, res) => {
+  try {
+    const rawMasterFile = await fs.readFileSync(masterFilePath, "utf-8");
+    const masterFile = JSON.parse(rawMasterFile);
+    res.json(masterFile[0]);
+    console.log("Successfully loaded Masterfile.");
+  } catch (err) {
+    console.error("Couldn't load Masterfile.");
+    console.error(err);
+  }
+});
+
 const documentsRoute = require("./routes/documents");
+const exportRoute = require("./routes/export");
+const versionsRoute = require("./routes/versions");
 
 app.use(express.json());
 app.use(express.static(rootPath));
-app.use(exportRoute);
 app.use(documentsRoute);
+app.use(exportRoute);
+app.use(versionsRoute);
 
 app.get("/", (req, res) => {
   res.sendFile(path.join(rootPath, "index.html"));